+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ### Latest/Nightly (this branch)
+- Features
+    - get_user_messages now returns the message index in addition to the frame index for each message.